--- apps/desktop/src/renderer/screens/main/components/Layout/index.ts
@@ -1,2 +1,0 @@
-export { MainLayout } from "./MainLayout";
-
